+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* pkg: [golang](https://golang.org/): `sync` and `update`
+
 * pkg: vscode: macOS: added a work-around to fix app auto-update
 
 ### Fixed